Add functionality
Added the functionality to color in the nodes as you mouse over them.
Added a reset colors button so that we're able to reset the grid without
resetting the rows and columns.

Unable to figure out how to account for a gap which is why we don't
include a gap.